+++ src/components/Balance/Balance.jsx
@@ -1,3 +1,4 @@
+import { RotatingLines } from "react-loader-spinner";
 import styles from "./Balance.module.scss";
 
 const Balance = ({ balance, isLoading }) => {
@@ -7,7 +8,11 @@ const Balance = ({ balance, isLoading }) => {
         <p className={styles.balanceClass__layout__description}>your balance</p>
         <h3 className={styles.balanceClass__layout__content}>
           {/* TODO: Add some spinner */}
-          &euro; {isLoading ? "Loading..." : balance}
+          {isLoading ? (
+            <RotatingLines strokeColor="black" width="30" />
+          ) : (
+            `€ ${balance}`
+          )}
         </h3>
       </div>
     </div>
