+++ src/components/Header/Header.jsx
@@ -1,10 +1,7 @@
-import { useNavigate } from "react-router";
-import authApiSlice from "../../redux/slices/api/auth/authApiSlice";
 import { useSelector } from "react-redux";
 import styles from "./Header.module.scss";
 import { ReactComponent as Logo } from "./../../img/logowallet.svg";
 import { ReactComponent as ExitDoor } from "./../../img/exitdoor.svg";
-import { RotatingLines } from "react-loader-spinner";
 import LogoutModal from "../LogoutModal/LogoutModal";
 import { Link } from "react-router-dom";
 import { useDispatch } from "react-redux";
@@ -14,24 +11,8 @@ import { openModal } from "../../redux/slices/modal/modalSlice";
 // import logoText from './../../img/logowallettext.png'
 
 const Header = () => {
-  const navigate = useNavigate();
   const user = useSelector((state) => state.auth);
   const dispatch = useDispatch();
-
-  const [logout, { isLoading, error }] = authApiSlice.useLogoutMutation();
-
-  const handleLogout = async () => {
-    const result = await logout();
-    if ("error" in result) {
-      //TODO: display in toast or something
-      console.log(error);
-    } else {
-      //TODO: display in toast or something
-      console.log("logged out successful");
-      navigate("/login");
-      navigate(0);
-    }
-  };
 
   return (
     <div className={styles.header}>
@@ -48,17 +29,13 @@ const Header = () => {
         <button
           className={styles.header__nav__exit}
           onClick={() => dispatch(openModal("logout"))}
-          disabled={isLoading}
         >
           <ExitDoor className={styles.header__nav__exit__icon} />
-          {isLoading ? (
-            <RotatingLines strokeColor="#bdbdbd " width="20" />
-          ) : (
-            <p className={styles.header__nav__exit__txt}>Exit</p>
-          )}
+
+          <p className={styles.header__nav__exit__txt}>Exit</p>
         </button>
       </div>
-      <LogoutModal logout={handleLogout} />
+      <LogoutModal />
     </div>
   );
 };
